Поставил функцию как в шаблоне

+++ Java/Level1.java
@@ -1,15 +1,15 @@
 import java.util.*;
 public class Level1
 {
-    public static long squirrel(int N)
+    public static int squirrel(int N)
     {
-        long result = 1;
-        long sum = 0;
+        int result = 1;
+        int sum = 0;
         for (int i = 1; i <= N; i++)
         {
             result *= i;
         }
-        for (long i = result; i > 0; i = i / 10)
+        for (int i = result; i > 0; i = i / 10)
         {
             sum = i % 10;
         }
